requirements.txt
appdirs==1.4.3
conekta==2.3.2
Django==1.10.5
django-annoying==0.10.3
django-paypal==0.4.1
django-tinymce==2.6.0
flake8==3.2.1
gunicorn==19.6.0
html5lib==1.0b10
httplib2==0.10.3
mccabe==0.5.3
nose==1.3.7
olefile==0.44
packaging==16.8
Pillow==4.0.0
psycopg2==2.6.2
py==1.4.33
pycodestyle==2.2.0
pyflakes==1.3.0
pyparsing==2.2.0
PyPDF2==1.26.0
python-dateutil==2.6.0
reportlab==3.4.0
simplejson==3.10.0
six==1.10.0
sorl-thumbnail==12.3
webencodings==0.5.1
whitenoise==3.3.0
xhtml2pdf==0.2b1
